sendmail problems fixed
